pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hmda-data-manager"
version = "0.1.0"
description = "Tools for managing CFPB's Home Mortgage Disclosure Act (HMDA) data for research"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "polars>=1.26.0",
    "pyarrow>=19.0.1",
    "python-decouple>=3.8",
    "scipy>=1.15.2",
    "selenium>=4.32.0",
    "webdriver-manager>=4.0.2",
    "requests>=2.31.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/hmda_data_manager"]
